HOCS-2239: add downscaler annotation (#86)
Presently, when the deployment changes the downscaler annotation
(that was manually added up to present) gets removed off the
outbound proxy. This causes the pod to stay up indefinitely, which
leads to increased costs and potential issues. This change enforces
the uptime meta-data annotation that enforces this pod to stay up.
This is specified differently depending on if we are in prod or
notprod.